Deflake color-resolving-hwb.html
This file adds an excessive 15552 tests, causing timeouts on
debug bots, even with timeout=long.

Instead of testing every possible permutation, it's probably enough
to test each parameter individually, using reasonable valid defaults
for the other parameters.

Additionally, move the "comma" tests out of the parameterized loops.
A few tests using known valid values (for the individual components)
should be enough to check that comma is not supported.

This reduces the number of tests to 36.

Fixed: 1282970
Change-Id: Id00eec083b8f5469d031e08b2e3bd393eedd9927