Fix iframe_sandbox_allow_top_navigation tests
The tests iframe_sandbox_allow_top_navigation-1.html and
iframe_sandbox_allow_top_navigation-3.html are failing because they are
using main frame navigations to data URLs to send a message to the
opener. But main frame navigations to data URLs were blocked by
https://codereview.chromium.org/2702503002.

So this CL change the tests to use a normal html file on the server
to send a message to the opener.

Bug: 853146

Change-Id: Ie6c8274d68bafb2683324cdd0c87881b7878c547