App history: implement finished promise
This also changes the abort reason used to signal abort on canceled navigations to be the same as the error passed to navigateerror and used for promise rejections. (This is tested by how the tests assert all errors are the same object.)

This contains a large overhaul of tests related to appHistory.transition and further refinement of the previously-introduced tests related to ordering, to be more exhaustive and to use a small event-recorder framework. It consolidates all tests about appHistory.transition into the ordering tests (with the directory now renamed ordering-and-transition). This is because both types of tests were setting up various situations and checking either the ordering of events (often not exhaustively), or the properties of appHistory.transition, or both.

Bug: 1183545
Change-Id: Idc914a98a4dd6579d003806a992b710881944770