Fix performance bug with scrollable areas in subtree layout
This change fixes the perf test in
third_party/blink/perf_tests/layout/subtree-layout-scrollable-area.html
by limiting the number of scrollable areas that are added to the frame
view. This regressed in https://crrev.com/1931148, which added all areas
to the frame view. This change improves performance in that test by 90%.

The previous behavior of only adding scrollable areas which are visible
to hit-testing and not `overflow: hidden` no longer works, since such
areas must still be considered for scroll anchoring. However, scroll
anchoring does not apply to areas whose content is shorter (in the block
flow direction) than the scrollable area, and so scrollable areas which
don't overflow in the block flow direction do not need to be added to
the frame view.

Bug: 1291551
Change-Id: I23270795b70d6a7d3b6f2c28cfc6e29d938f7943