[COEP] Improve worker COEP inheritance for `filesystem:` URLs.
This is not the ideal behavior - that would be to inherit the policy
from the execution context that created the file. That is blocked on
crbug.com/1146362.

In the meantime, this is a marked improvement over the current situation
where this bug allows bypassing COEP.

This change adds Web Platform Tests to cover the changed behavior, even
though shared workers are unaffected due to crbug.com/987517.

Bug: chromium:1291230
Change-Id: I2d944bd3302a7dd9ff57b9e7e71c535846cdbddc